match: break early when matching arg{,path}X
At most a filter can contain 64 entries of each, but in the common case
only the first few (if any) are used at all. Remember the largest offset
in each filter, and stop matching once it has been surpassed.

Also, move the checks that are very likely to succeed from the beginning
to the end of the function.

Signed-off-by: Tom Gundersen <teg@jklm.no>