unit: tell nss-systemd not to query the bus when called from dbus-broker
In case nss-systemd is triggered due to looking up a username from
a policy file (unlikely in a standard setup, but possible) it would
try to access the bus in order to answer the query.

However, PID1 and the nss module have a hack to avoid this: When PID1
discovers that it is starting dbus.serivce, it will add a special
environment variable which the NSS module will check for. In case it
is found, the NSS module will bypass the bus, avoiding the deadlock.
See <https://github.com/systemd/systemd/commit/fd63e712b2025d235ce4bfbb512fada10e2690b5>
for details.

We are started as dbus-broker.service (with dbus.service as an alias),
and as such PID1 was not adding the environment variable for us. This
adds it explicitly in our service file.

Signed-off-by: Tom Gundersen <teg@jklm.no>